backend/detectAlgorithm/requirements.txt
backend==0.2.4.1
matplotlib==3.6.2
mediapipe==0.9.1.0
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==1.5.2
PyYAML==6.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
skimage==0.0
tensorflow==2.12.0
tensorflow_hub==0.13.0
tensorflow_intel==2.12.0
ultralytics==8.0.49
xgboost==1.7.5
